Some refactoring and cleanup.  Got rid of some fixtures and move the migration code around for easier reuse.
